requirements.txt
flask
groq
requests
pillow
apscheduler
pywebpush
python-dotenv